refactor: move winding logic to separate module
